requirements.txt
glcontext==3.0.0
keyboard==0.13.5
mido==1.3.3
moderngl==5.12.0
noise==1.2.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==25.0
pillow==11.3.0
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
python-rtmidi==1.5.8
pywin32==311
PyYAML==6.0.2
sounddevice==0.5.1
fastapi==0.115.6
uvicorn==0.34.0
obs-websocket-py==1.0
setuptools==80.9.0
usb-monitor==1.23
wheel==0.45.1
WMI==1.5.1